+++ api/CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0] - 2020-12-16
+
+### Added
+
+- Add `ffi_derive` for `Measurement` and `Unit` (behind the `cffi` feature flag).
+
+### Changed
+
+- Updated `ffi_common` and `ffi_derive` to 0.2.
+- `wise_units-ffi`'s `build.rs` now uses `OUT_DIR` instead of
+    `CARGO_MANIFEST_DIR`.
+
+### Fixed
+
+- Pointed `agrian-registry` to GitHub.
+
 ## [0.15.0] - 2020-07-31
 
 ### Changed